# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prairie Connect API",
    "version": "0.1.0",
    "description": "API for the Prairie Connect rail logistics platform — directory/marketplace,\nuniversal search, corridor atlas, route finder, and the AI guide.\n\n**Auth**: integrate directly with the Cognito User Pool (Amplify Auth or\namazon-cognito-identity-js) and send the resulting JWT as a Bearer token.\n\n**Realtime**: fetch `/v1/realtime/token` for the MQTT-over-WSS config;\nchat replies stream there as `{type: delta|action|done}` events.\n\n**Chatbot UI actions**: assistant replies carry an `actions` array (see the\n`UiAction` schema) the client should execute — show a map, navigate, draw a\nroute, open directory results."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito JWT (id or access token). Sign in against the Cognito User Pool (see stack outputs userPoolId / userPoolClientId) and paste the token here."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "RealtimeConfig": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "description": "IoT WSS endpoint host"
          },
          "authorizer": {
            "type": "string",
            "description": "IoT custom authorizer name"
          },
          "topicPrefix": {
            "type": "string"
          },
          "topics": {
            "type": "object",
            "properties": {
              "notifications": {
                "type": "string"
              },
              "chatPattern": {
                "type": "string",
                "description": "Chat topic; replace {conversationId}"
              }
            },
            "required": [
              "notifications",
              "chatPattern"
            ]
          }
        },
        "required": [
          "endpoint",
          "authorizer",
          "topicPrefix",
          "topics"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/realtime/token": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "WebSocket connection config",
        "description": "Returns everything needed to open the MQTT-over-WSS connection: connect to `wss://<endpoint>/mqtt?x-amz-customauthorizer-name=<authorizer>` passing your Cognito JWT as the MQTT password (any username). Subscribe to the topics returned — the authorizer scopes you to your own subtree.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Realtime connection config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeConfig"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}